Move settings classes to their relevant files
