+++ client/background.ts
@@ -1,0 +1,29 @@
+import { h } from 'snabbdom/h';
+import { VNode } from 'snabbdom/vnode';
+
+import { _ } from './i18n';
+import { Settings } from './settings';
+import { radioList } from './view';
+
+const backgrounds = {
+    light: _("Light"),
+    dark: _("Dark"),
+};
+
+class BackgroundSettings extends Settings<string> {
+
+    constructor() {
+        super('theme', 'light');
+    }
+
+    update(): void {
+        document.documentElement.setAttribute('data-theme', this.value);
+    }
+
+    view(): VNode {
+        return h('div#settings-background', radioList(this, 'background', backgrounds, (_, key) => this.value = key));
+    }
+
+}
+
+export const backgroundSettings = new BackgroundSettings();
